simple_admin/requirements.txt
django==2.2.0
mysqlclient
djangorestframework
coreapi
Markdown
django-jsonfield
djangorestframework-datatables
paramiko
dwebsocket
django-filter
celery==4.3.0
ansible==2.7.10
redis==3.2.1